fix (admin page)

+++ src/Pages/appointmentSuccess.jsx
@@ -175,7 +175,7 @@ const Success = () => {
 
 export default Success;
 export const formatDate = (dateString) => {
-  const options = { year: "numeric", month: "long", day: "numeric" };
+  const options = { year: "numeric", month: "short", day: "numeric" };
   const date = new Date(dateString);
   return (
     date.toLocaleDateString(undefined, options) +
